fix: link collector error handling (#18)
I noticed in `LinkCollector` there didn't appear to be error handling for
the case that the `ActorSource.actorRef` would fail (when hitting the
5000 buffer size), plus the termination condition of
`.idleTimeout(500.millis)` seemed like a workaround.

I tried to, instead of using a stream here, use regular recursion. This
perhaps reduces the parallelism in reading the site from disk, but the
bottleneck is likely in the network calls anyway, and those are still
in parallel to reading from disk.